requirements.txt
# requirements.txt
Flask>=2.0 
Flask-Cors>=3.0
requests>=2.0
python-dotenv